find: use the plugins delete callback for -delete
This patch changes the `-delete` handling of rbh-find to instead use the
callback from the plugins and extensions. Each plugin's callback is
called, but the first one to send back a success rc ends the loop.

The unlink is kept specifically for the MPI plugin, until the proper
callbacks are implemented.

+++ rbh-find/src/find_cb.c
@@ -264,7 +264,27 @@ find_exec_action(struct find_context *ctx,
 {
     switch (action) {
     case ACT_DELETE:
-        return unlink(fsentry_relative_path(fsentry));
+        /* XXX: will be changed once all plugins handle delete properly */
+        if (ctx->info_plugin_count != 0 || ctx->info_extension_count != 0) {
+            int rc;
+
+            for (int i = 0; i < ctx->info_extension_count; ++i) {
+                rc = rbh_extension_delete_entry(ctx->info_extensions[i],
+                                                fsentry);
+                if (rc == 0)
+                    return rc;
+            }
+
+            for (int i = 0; i < ctx->info_plugin_count; ++i) {
+                rc = rbh_plugin_delete_entry(ctx->info_plugins[i], fsentry);
+                if (rc == 0)
+                    return rc;
+            }
+
+            return rc;
+        } else {
+            return unlink(fsentry_relative_path(fsentry));
+        }
     case ACT_EXEC:
         return exec_command(ctx, fsentry);
     case ACT_PRINT:
